Support async OpenSearch and use separate method to get connection
- Use opensearch-py[async]. By itself this is pointless because the
logging interface is synchronous, but it's useful when opensearch-based
logging is used in an async project which also uses opensearch for other
things.

- Remove the connection parameters and instead take a method which
returns a client. This is helpful in scenarios where the client must be
re-initialised with new credentials periodically, as the external
function can manage that lifecycle.

- Remove the test_opensearch_connection method as all it did was call
ping() on the client, and if we're building the client externally we can
do that ourselves (asynchronously).